change error handling and logging logic
Make beautiful error handling on frontend.
Add helper to log and send error for http handlers.
Change slog config to add source of call.
Change database driver in error handling because previous it is not the one sqlc use, so error comming from sqlc was not understadable by parseError function. Now it is work correctly.